[fe] refactor (#61) 디렉토리 이동
